\name{LncRIndiv}
\alias{LncRIndiv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Individual-level differential expression detecting method
}
\description{
To detect differentially expressed lncRNAs at the individual level.
}
\details{
\tabular{ll}{
Package: \tab LncRIndiv\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-12-2\cr
License: \tab LncRIndiv\cr
}

For usage, see the documentation of the main function 'LncRIndiv'.
}